fix: write the validated filtered file
